fortysixth/requirements.txt
acme==1.1.0
appdirs==1.4.3
asgiref==3.3.4
attrs==19.3.0
Automat==0.8.0
blinker==1.4
certbot==0.40.0
certbot-nginx==0.40.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloud-init==20.2
colorama==0.4.3
command-not-found==0.3
ConfigArgParse==0.13.0
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.2
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-quill==0.1.0
django-quill-editor==0.1.22
docopt==0.6.2
entrypoints==0.3
filelock==3.0.12
future==0.18.2
gevent==1.4.0
greenlet==0.4.15
gunicorn==20.0.4
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
Jinja2==2.10.1
josepy==1.2.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
loguru==0.5.3
MarkupSafe==1.1.0
mock==3.0.5
more-itertools==4.2.0
netifaces==0.10.4
oauthlib==3.1.0
parsedatetime==2.4
pbr==5.4.5
Pillow==8.2.0
pipreqs==0.4.10
psycopg2==2.8.4
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyICU==2.4.2
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.3
python-debian===0.1.36ubuntu1
python-dotenv==0.17.1
pytz==2019.3
PyYAML==5.3.1
requests==2.22.0
requests-toolbelt==0.8.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sqlparse==0.2.4
ssh-import-id==5.10
systemd-python==234
Twisted==18.9.0
ubuntu-advantage-tools==20.3
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
virtualenv==20.0.17
wadllib==1.3.3
yarg==0.1.9
zipp==1.0.0
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.0.0
zope.interface==4.7.1
